Migrate search route to Axum
