+++ dim-web/src/routes/mod.rs
@@ -2,5 +2,6 @@ pub mod auth;
 pub mod dashboard;
 pub mod filebrowser;
 pub mod library;
+pub mod search;
 pub mod tv;
 pub mod websocket;
